Created New Class
Created a Point Class to show how a method and an override for the method can be created.  The overload here creates two different signatures of the method.  The first with 2 integer values and the other overlod with a single Point value.

Shortcut:  {SHIFT} + {ALT} + C
Description:  Creates a new Class Object