New Class And Method
Class name is Calculator and the method is Add.  The trick here is to use the params keyword which allows you to simplify your code and pass a list of numbers or a new integer array.    This can simplify your code.

Shortcut:  {CTRL} + {SHIFT} + C
Description:  New Class

Snippet:  foreach
Description:  Creates a new Foreach Loop