Logical Error - Need Access Modifier
Defensive Coding practices can prevent this kind of error from occurring.  In this case we create two orders and the count should be 2.  Then we promote and the orders list is reset and the count returns to 0.  In the next change you will see how the readonly modifier can be used to resolve this issue.