Reorganizing Person.cs
Placing Properties at the top with constructors followed by and Methods or Functions.

Methods tend to be used generically however I want to differentiate between a Method and a Function.  Method does not have a return Value and a Function does.  However, most programmers use the words method and function interchangeably