Created New Customer Class
In addition to creating the new Customer.cs I noticed that after I renamed the class Customer_old that I still needed to rename the file.

This new class was created to show the need to keep business logic that is related to a class private and not accessible to other classes.  This is the most commonly used access modifier.