Adding Reference to CafeDeLeMente
Reference Has Been Added to CafeDeLeMente in the project CSharpBasics.  Additionally was required to make the 3 classes public in order to be accessible to other projects.  Classes are assumed to be private unless an access modifier is added.  In this case we made the access modifiers public so we could access the Customer class from the Main Program.cs.